requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3